created helper class to help calculating prices
