+++ src/main/java/nl/garagemeijer/salesapi/dtos/sales/SaleInputDto.java
@@ -21,6 +21,8 @@ public class SaleInputDto {
     private String paymentMethod;
     @NotNull(message = "please fill in if the vehicle is for business or private use")
     private String businessOrPrivate;
+    @NotNull(message = "please fill in what type this is, offerte or order")
+    private String orderType;
     private String comment;
     private Addition addition;
